refactor: 식별자 값이 존재하지 않는 Line 객체가 존재하지 않도록 생성자를 변경
